+++ src/django_mysql/models/fields/__init__.py
@@ -1,7 +1,6 @@
 from django_mysql.models.fields.bit import Bit1BooleanField, NullBit1BooleanField
 from django_mysql.models.fields.dynamic import DynamicField
 from django_mysql.models.fields.enum import EnumField
-from django_mysql.models.fields.json import JSONField
 from django_mysql.models.fields.lists import ListCharField, ListTextField
 from django_mysql.models.fields.sets import SetCharField, SetTextField
 from django_mysql.models.fields.sizes import SizedBinaryField, SizedTextField
@@ -10,7 +9,6 @@ __all__ = [
     "Bit1BooleanField",
     "DynamicField",
     "EnumField",
-    "JSONField",
     "ListCharField",
     "ListTextField",
     "NullBit1BooleanField",