+++ tests/testapp/test_forms.py
@@ -2,9 +2,8 @@ import pytest
 from django import forms
 from django.core import exceptions
 from django.test import SimpleTestCase
-from django.utils.html import escape
 
-from django_mysql.forms import JSONField, SimpleListField, SimpleSetField
+from django_mysql.forms import SimpleListField, SimpleSetField
 
 
 class TestSimpleListField(SimpleTestCase):
@@ -243,79 +242,3 @@ class TestSimpleSetField(SimpleTestCase):
         with pytest.raises(exceptions.ValidationError) as excinfo:
             field.clean("")
         assert excinfo.value.messages[0] == "This field is required."
-
-
-class TestJSONField(SimpleTestCase):
-    def test_valid(self):
-        field = JSONField()
-        value = field.clean('{"a": "b"}')
-        assert value == {"a": "b"}
-
-    def test_valid_empty(self):
-        field = JSONField(required=False)
-        value = field.clean("")
-        assert value is None
-
-    def test_invalid(self):
-        field = JSONField()
-        with pytest.raises(exceptions.ValidationError) as excinfo:
-            field.clean("{some badly formed: json}")
-        assert (
-            excinfo.value.messages[0]
-            == "'{some badly formed: json}' value must be valid JSON."
-        )
-
-    def test_prepare_value(self):
-        field = JSONField()
-        assert field.prepare_value({"a": "b"}) == '{"a": "b"}'
-        assert field.prepare_value(["a", "b"]) == '["a", "b"]'
-        assert field.prepare_value(["你好，世界", "😀🐱"]) == '["你好，世界", "😀🐱"]'
-        assert field.prepare_value(True) == "true"
-        assert field.prepare_value(False) == "false"
-        assert field.prepare_value(3.14) == "3.14"
-        assert field.prepare_value(None) == "null"
-        assert field.prepare_value("foo") == '"foo"'
-        assert field.prepare_value("你好，世界") == '"你好，世界"'
-        assert field.prepare_value("😀🐱") == '"😀🐱"'
-
-    def test_redisplay_wrong_input(self):
-        """
-        When displaying a bound form (typically due to invalid input), the form
-        should not overquote JSONField inputs.
-        """
-
-        class JsonForm(forms.Form):
-            name = forms.CharField(max_length=2)
-            jfield = JSONField()
-
-        # JSONField input is fine, name is too long
-        form = JsonForm({"name": "xyz", "jfield": '["foo"]'})
-        assert "[&quot;foo&quot;]</textarea>" in form.as_p()
-
-        # This time, the JSONField input is wrong
-        form = JsonForm({"name": "xy", "jfield": '{"foo"}'})
-        # Appears once in the textarea and once in the error message
-        assert form.as_p().count(escape('{"foo"}')) == 2
-
-    def test_already_converted_value(self):
-        field = JSONField(required=False)
-        tests = [
-            '["a", "b", "c"]',
-            '{"a": 1, "b": 2}',
-            "1",
-            "1.5",
-            '"foo"',
-            "true",
-            "false",
-            "null",
-        ]
-        for json_string in tests:
-            val = field.clean(json_string)
-            assert field.clean(val) == val
-
-    def test_disabled(self):
-        class JsonForm(forms.Form):
-            jfield = JSONField(disabled=True)
-
-        form = JsonForm({"jfield": '["bar"]'}, initial={"jfield": ["foo"]})
-        assert "[&quot;foo&quot;]</textarea>" in form.as_p()